additional diagnostics: ratio of ndt to RT
I added this to get a simple measure for whether the model thinks that the
subject used the information provided by the shown dots. This seems to work. For
example, subject 3 has median ratio of 0.29 whereas subject 6 has median ratio
of 0.66. So, although the posterior predictive log likelihoods are worse for
subject 3, the median ndt to rt ratio shows that his or her responses may follow
the dots much more closely than those of subject 6.