added collapsing bounds
I tested subjects 3, 4 and 6 with and without collapsing bounds and did not find
strong advantages of collapsing bounds. For subject 3 there may be a small
advantage. For the others, for which the large ratio of ndt to total RT
also indicates some difficulties in fitting, collapsing bounds appear to lead to
smaller ppls for test trials, i.e., worse predictions. In any case, the bstretch
parameter was usually around 0.2 also indicating only limited collapse.

Using collapsing bounds seems to decrease the noisestd, though. In subject 3,
this was offset against an increase in ndt and its variability. Hence there may
have not been any noise reduction in the end. In subject 4, noisestd was reduced
without increases in ndt and its variability indicating overall less amount of
noise in the model, but at the cost of smaller ppls on the test trials, i.e.,
slightly worse predictions.