new script for parameter inference for all subjects
infer_behaviour.py runs EP-ABC for each subject of the BeeMEG experiment, or
selected subjects, and for each specified model (currently either basic or
collapse). It uses helpers.py to load the data.

EP-ABC uses the new parameter container.

Exceptions are caught to prevent the script to break without trying to infer
parameters for all subjects. The script will try to run EP-ABC 3 times before
skipping the subject.

All relevant settings and results are stored in a numpy-compressed file archive.

The settings committed here are for testing purposes (only 2 subjects and
minacc=200 instead of 2000).