new script collecting first analyses of inference results
Will load results from inference with the basic and the collapsing bound model
and subject responses and then provide the following analyses:
- compare basic and collapsing bound models across subjects for a chosen measure
- show relations between different measures from one model across subjects
- check whether there are any obvious sequential effects in subject's response
  times

So far I have run inference on the complete data set twice, as indicated by the
two sets of file names. In the earlier set there was an error for subject 15 in
the collapsing bound model. I did not catch exceptions when this happened. So
there's no data for this subject. I could not reproduce this error and in the
later set there are no missing subjects anymore, but some -inf ppls which I
still need to fix.