new analysis of single subject responses (e.g. why high logml?)
To investigate possible causes or interpretations of the summary measures for
individual subjects I now check the response distribution of a selected subject.
I use two different visualisations: standard response distribution plots in
which trials are pooled and a plot where responses of all trials are shown
individually. In the latter I sort trials according to choice and RT and then
plot them in that order together with a sample from the model. Ideally I wanted
to see that the model predictions follow somehow the shape of the subject
responses, but apart from a general tendency that the model makes the same
choice as the subject, I didn't recognise much structure. Therefore, the plot
provides just an intuition for how uncertain the model predictions (and/or
subject response) still are after inference.

NOTE: I noticed that the two inference results currently available are very
different. I currently don't see any reason for this and speculate that one of
them had interferring variables in the workspace when I started inference. The
first set appears to be more suspicious, especially because it's the one with
the unreproducable error. So I stick to the second set for now, awaiting results
from the currently running set.

concrete checks so far:

subject 32 - only one with logml>-300 for collapsing bound model
----------------------------------------------------------------
I originally thought that the small logml would inidicate that we can predict
the subject's responses particularly well, but I didn't find any immediate
indicators of that. The sum of ppls is almost perfectly correlated with the
logml so that it does not give an independent measure of how well the responses
can be predicted. Is that surprising? Probably not. The sum of ppls for test
trials is more informative, but it also correlates strongly with the logml. This
is a good thing as it means that the logml of the training trials gives a good
indicator for how well responses in test trials can be predicted by the model.
In that sense the high logml does actually show that we can predict better with
the corresponding model. I just didn't see this in the single trial plot for
subject 32 where the model predictions were still very noisy.

A worry: is that there is a strong anti-correlation between logmls and medianRT,
i.e., the smaller the medianRT, the higher the logml (pearsonr=-0.76). One
reason for this could be that the RT standard deviations are also smaller with
small medianRT meaning that there is simply less variance in the data that the
model needs to explain. Indeed, there's also a strong anti-correlation between
RT std and logml (pearsonr=-0.64), but this is smaller than the anti-correlation
with medianRT suggesting that medianRT may be a more important factor.

subject 8 - unusual combination of logml and RT std
---------------------------------------------------
The lower anti-correlation of RT std mainly comes from a subject (8) which had a
relatively high logml together with a very high RT std. This combination is a
good sign, because it suggests that the model can explain relatively much of the
RT variance. It also does so for this subject by using only ndt with a mode of
0.34 vs. dtmode of 1.1. Yet, the single trial plot still does not show much
recognisable structure.

The anti-correlation between logml and medianRT may also suggest that the chosen
prior strongly favours small RTs. This assumes that the logml considerably
reflects the prior. It's worth checking out further.

subject 23 - mediocre logml, but reasonable fit?
------------------------------------------------
To contrast my findings for the subject with the logml I checked a subject with
mediocre logml of about -500 (subject 23). I was surprised that other indicators
of a good fit actually appeared to look better than those for subject 32.
Specifically, the noisestd was lower (49 vs. 59) and ndt_vs_RT was lower (0.40
vs. 0.51). This appears to be mainly due to the generally higher RTs of subject
23 which the model explains with dt instead of ndt. That's good, but also lead
to lower logmls. Why?

I should be careful not to overinterpret these logml differences between
subjects. After all their responses were very different. It may be true, though,
that the model currently favours fast responses.