analysis of new results + parameter checks + delete first results
delete first results
--------------------
The new results (set 3) have confirmed that the first results (set 1) were some
strange outlier. I plotted the logmls of all results sets together in one plot.
This showed that sets 2 and 3 had very similar logmls while set 1 shared the
high values of logmls, but had considerably lower logmls for some subjects with
medium or low logmls, especially subjects 2, 5, 8, 12, 13, 17, 18, 21, 25, 26,
27, 28, 29, 30, 36. It's still unclear to me why this happened. Mostly it's a
mystery, because the subjects with deviating logmls appear to have random order.

Having confirmed that something went wrong in the first run of inference, I
deleted these results.

analysis of new results
-----------------------
In the remaining results (sets 2 and 3) I could not find a clear distinction
between subjects for which inference appears to have worked and for which this
is not the case. Specifically, there is no visual clustering of subjects
according to the selected measures (accuracy, stdRT, medianRT, ndt_vs_RT, logml,
pplsum_test, noisestd) in the pairplot.

The best that came to my mind was to discriminate based on medianRT (set as hue
now), but this only shows previous insights: people with low medianRT tend to
have low accuracy and high logml.

As the other results set, set 3 also shows a clear advantage for collapsing
bounds. Only one subject had higher logml for the basic model. Especially
subjects with high logmls show big advantages of collapsing bounds. I get a very
similar picture when using the pplsum over test trials as measure. With
pplsum_test this is a bit more noisy and 4 subjects show an advantage for the
basic model, but the overall trend is the same. pearsonr between logml and
pplsum_test is 0.92.

parameter checks
----------------
Investigating inference results further I started to check parameter posteriors.
I was surprised to find that these look very reasonable for subjects 23 and 32
despite their large difference in logml. In particular, the posteriors are
quite narrow and mostly have only small correlations below 0.4. Parameters with
larger correlations in subject 32:

bshape + bstretch,
prior + lapsetoprob,
bound + bstretch,
noisestd + bound,
noisestd + ndtspread;

of these noisestd+ndtspread was the most severe with -0.59, but as the
posteriors were generally narrow, this does not worry me. Parameters with larger
correlations in subject 23:

noisestd + ndtmean
bstretch + ndtmean
lapsetoprob + bound

of which bstretch+ndtmean was the largest with 0.7. Also there bstretch only
ranges from 0.32 to 0.44 which appears small and ndtmean ranges from -0.84 to
-0.94 which is also small.

Finally, I checked the fitted posterior for the subject with the worst logml
(subject 17 with -676). The posterior for this subject is generally wider and
has larger correlations between parameters with many of them between 0.2 and
0.5. The three largest are:

bshape + bstretch (-0.64)
noisestd + bound (0.6)
ndtmean + bstretch (0.52)

A striking thing about this result is that the prior is very low (~0.32). I
checked whether it is generally the case that with strong prior of choice
towards one alternative the logml is low. There is a correlation across subjects
suggesting that, but it's not so strong (pearsonr=0.41).