entropies of posterior response distributions for all trials
New script estimates and stores the entropies of the response distributions
predicted by the posterior for each subject and trial for a given stored
inference result.

first analysis (infres_collapse_201612121759_entropies.npz):

relation to ppls
----------------
Summed over trials, posterior predictive log-likelihoods and entropies are
strongly anti-correlated across subjects (pearsonr=-0.94). This suggests that
these roughly measure the same thing when averaged over trials.

This is different on the level of single trials. Although there is still an
anti-correlation between the two, it is much weaker (pearsonr=-0.37). I guess
that this is because ppls depend on the responses of the subject which are noisy
while the model entropies summarise the whole response distribution predicted by
the model.

structure in entropies across trials
------------------------------------
Plotting the entropies for single subjects across trials reveals a clear
structure which appears to be determined by the particular trials that we chose
for the experiment. Specifically, the trials are ordered according to the 5th
dot manipulation such that 28 trials with the same 5th dot location are lumped
together. There were 6 different 5th dot locations embedded in otherwise the
same 28 dot location sequences. This is followed by 72 "catch trials" which were
trials without specific dot location sequences, but which had generally faster
responses. This whole trial sequence was repeated (with mirrored x-locations) to
get the total 480 trials.

This structure can be clearly seen in the entropies: There are first 6 blocks of
trials where the middle blocks (corresponding to 5th dot locations close to the
centre of the screen) show higher entropies than the first and last blocks
(having extreme 5th dot locations). These are followed by a block of trials
showing no clear structure, but having relatively low entropies, probably
because predicted responses are compact and fast. Then the same structure is
repeated once.

All the subjects showed this structure more or less in their entropies although
at different levels of entropy; some with a mean entropy around 3.0 (e.g. sub
33) others with a mean entropy around 3.7 (e.g. sub 25). I guess it shouldn't be
surprising that this structure is present in models after inference. After all
we put it there by making the model depend on the dot locations. Perhaps the
only thing it shows us is that the model did not need to overwrite all of the
dot information with noise to explain the subject responses.

The entropies, or rather the structure in the entropies, may then be a measure
for identifying subjects which should exhibit a clear effect of the 5th dot
manipulation. For that I would have to develop a measure for how strongly our
trial structure is reflected in the entropies. Alternatively, this may just be
the same as looking for subjects with small noisestd and ndt.