First basic version of analysis looking for evidence in behaviour
that the 5th dot influenced the choices of a subject. The basic idea is to see
whether 5th dot positions that support the correct choice (according to the
first 4 dot positions) more also lead to more correct responses.

To do this in combination for choice and RT I define a "penalty" that penalises
fast wrong choices more than slow correct choices.