added analysis of binned fraction correct and median RT
For the tested subject (2) there are very clear effects in both fraction correct
and median RT across the binned and flipped 5th dot positions.

The median RT decreases as support for the correct choice increases, as
expected. Interestingly, this relation reverses for the most extreme bins and
their neighbours, i.e., the bin with the least support for the correct decision
has smaller RT than the next bin and the bin with the most support for the
correct decision has larger RT than the previous bin. Does this indicate that
people ignore these extreme dot positions? I guess I should check whether this
is consistent across people.