logistic regression of whether response was correct onto support_correct
Very simple with statsmodels and pandas. All subjects I looked at so far showed
a strong dependence of their responses on how much the 5th dot supported the
correct response. For the two subjects I looked at the p-value for the null of
the logistic regression with only the intercept was below 1e-10.