ran logistic regression on choices across all subjects
As my initial tests with single subjects indicated, there appears to be a clear
effect in all subjects. Even the subject with the highest p-value of 2.15e-06
exhibits a clear increase of fraction correct choices with increasing support
for correct choice bin.

I also see effects in RT, but they don't appear to be as strong, or easily
interpretable as the effects in choice.

Given that we see this strong effect in choice, does that mean that subjects
made a response immediately after the 5th dot in many trials? Not necessarily,
as the 5th dot may have influenced also later choices.