basic regression produces significant but unreliable results
To further investigate the significant effects in Hame's epoch data I first
adapted the visualisation so that the range of the colormap is the same across
subjects. I also found an option for which automatically the 7 highest peaks in,
e.g., negative log p-value are shown. Checking across 4 subjects peaks appeared
to occur at roughly similar times and reached p-values of 0.0001, but some of
the peaks were also before first dot presentation time (negative times)
indicating that there was a general issue.

To check MNE's regression procedure I ran a linear regression manually with
statsmodels for the data at one of the peaks. The raw data also showed a
significant (0.0003), but small (0.16) correlation in a seaborn jointplot
between MEG signal (mag) and how much the 5th dot supported the previous
evidence in a trial. This correlation was reduced, but still significant when
I set potential outlying values to 0. The statsmodels regression confirmed a
significant effect of support_correct with a p-value of 0.0006. So there is
genuinely something in the data which produces significant regression results.

Triggered by some p-value peaks being even before first dot onset I finally
checked in how far these results actually depend on the support_correct values
and reran the analysis with a randomly permuted set of support_correct values.
The results confirm that the significant regression results are independent of
the particular support_correct values used in the analysis, at least regarding
the mere occurrence of "highly significant" peaks in the data (obviously the
position of the peaks depends on the permuted order of support_correct values).
The basic regression analysis I ran, therefore, is not reliable in the sense
that p-values violate their interpretation of excluding random results.
Consequently, the linear regression must be the wrong model for the data and I
have to think about something else.