checking cluster-level permutation p-values for basic regression
MNE provides non-parametric, permutation-based inference on the cluster-level.
The procedure works as follows:

1) identify clusters based on a user-defined test statistic and given a
   connectivity matrix for the given variables
2) compute a cluster-level test statistic by summing the values of the test
   statistic for all variables in the cluster
3) repeat many times:
    3.1) randomly permute the MEG-data values
    3.2) re-compute the test statistic
    3.3) re-compute the cluster-level test statistic
4) compute the p-value as the fraction of times the original cluster-level test
   statistic fell below the permuted cluster-level test statistic

I decided to use MNE's spatio_temporal_cluster_1samp_test which is suited for my
case of a single predictor (support_correct). My test statistic was the t-value
of the corresponding regressor. To get results reasonably fast I had to
implement a numba-jitted function which can compute these t-values very quickly
for a large number of variables (see helpers.linregress_t).

The results confirm my previous observation that the identified clusters are
unreliable: For the 17 clusters identified with a t-value based threshold of 3
the lowest cluster-level p-value was 0.51 with most of the p-values approaching
1.

It is a bit strange to observe that almost all random permutations lead to
larger t-value sums on the cluster level. Why should this happen? Even for
completely random data you would expect that in half of all cases the t-values
of the oringial clusters are larger than the permutations! Perhaps this is
related to how MNE permutes the data: It randomly flips the sign of the sensor
values. I should investigate this further.