basic permutation verifies p-values of basic regression
Permuting the MEG values of a given sensor at a given time across trials and
then recomputing the t-values for support_correct of 5th dot leads to
the p-values of the original linear regression. Perhaps this should not be
surprising.

I originally did this to understand why permutation-based cluster-level p-values
tended to be close to 1. Maybe this is then related to summing t-values across
sensors in a cluster.