bugfix + cluster-level permutation statistic works differently
bugfix: did not use chname of found peak in manual regression analysis, but
        hard-coded channel name

I misunderstood how the cluster-level permutation statistic works. They do not
re-compute the sum of t-values for the identified clusters under permuted data,
but re-compute clusters completely for the permuted data and then compare
against the maximum of the t-value sums across the clusters found for the
permutation. So each permutation has exactly one summed t-statistic and the
summed t-statistic for an identified cluster has to be larger than that of the
permutation to reduce the p-value. A p-value close to 1 means that the summed
t-statistics of the largest/most significant cluster of (almost) all
permutations were bigger than the summed t-statistic of the identified cluster.

The distribution of maximum summed t-statistics across permutations (H0) covers
a large range of values which, surprisingly, exceeds the summed t-statistics of
the originally identified clusters considerably, hence, leading to the high
p-values > 0.5. I find it strange that random sign-flips of the MEG-data lead to
so much bigger/more significant clusters, but it is unclear whether this is due
to many more sensor-time-pairs being in a cluster, or due to much higher
t-values.