random large clusters + influence of threshold and max_step
I traced the origin of the large summed t-statistics and confirmed that these
result from large clusters, not extraordinary large t-statistics. Specifically,
the average t-statistic in a random cluster is mostly between 3.0 and 3.5, but
the number of variables (across space and time) contributing to a cluster may go
beyond 20 while there were only 4 clusters in the genuine data with more than 10
variables (12, 13, 14 and 16).

I then investigated two potential remedies for this situation: 1) increasing
threshold and 2) increasing max_step.

threshold
---------
If t-values up to 3.5 are very likely for my data, it could be that highly
significant clusters that I have are masked by random large clusters with only
moderately high t-values. To exclude this possibility from the analysis I
increased the cluster-forming threshold to 4.0. Then, only 2 clusters are left
in the original data and their permuted cluster-level p-values drop, but only to
about 0.2, because there are still sufficiently many larger random clusters.

max_step
--------
This parameter supposedly controls how many consecutive time steps are treated
as neighbours during clustering. I here set this to 5 under the assumption that
all time steps up to 20ms apart may be neighbours. I'm not sure whether this
setting makes sense. Does it mean that time step 1 may be a neighbour to time
step 5 independent of the value of the variable at time step 2? If yes and if
the neighbourhood relationships are transitive during the implemented analysis
(a neighbour of my neighbour is also my neighbour), setting max_step to
something different than 1 wouldn't make sense.

It anyway doesn't appear to have an effect. As far as I can tell, the same
clusters are found with max_step=5 as with max_step=1.