discrepancies between t-values due to projection
I stumbled upon a weird discrepancy between the t-values I computed and the
t-values that MNE computed. It took me a long time to figure out that MNE
automatically introduced some 'projection' to the data when I put the data in an
epochs container. This meant that the data returned by the epochs container was
different from the data that I put in.

I now prevent the application of the projection by setting proj=False when
creating the container. This makes all t-value computations consistent.

Without the projection t-values are a bit lower and the cluster-level p-values
are even higher. Perhaps because the number of variables in the clusters reduced
slightly (highest numbers: 13, 12, 9, 7) although the number of clusters
remained at 17.