properly permute data instead of sign-flipping: even larger p-values
NOTE: I made the changes for the analysis directly in the MNE code in
mne-python/mne/stats/cluster_level.py in function '_do_1samp_permutations'.
See my commit '[master 691156f]' there.

My motivation was that the random sign-flip permutation may be right for a test
of the null that the data has a mean of 0, but this certainly does not apply for
my regression analysis where the test statistic investigates a slope while
accounting for the intercept (mean). This does not mean, though, that random
sign flips cannot destroy the relationship with my predictor variable as they
will also increase large fluctuations in slope, but the sign flipping may
exaggerate properties of the data (like an intercept/mean/offset) that mask the
influence of the predictor variable.

It turns out that this appears to be true. With proper permutations the randomly
generated clusters based on the t-statistic of the support_correct regressor
become even more significant/larger, leading to even larger p-values of the
identified clusters in the data with the lowest p-value being 0.781. This makes
my analysis even more questionable. The distribution of maximum summed
t-statistics over permutations even indicates that its modes take on larger
values than the values from the data clusters. Why do random permutations of the
data produce bigger/more significant clusters than the data itself?