run cluster-level permutation analysis on all subjects
I still wondered how it could be that almost 80% of maximum clusters identified
in permuted data were bigger than the largest cluster in the original data. But
this can simply happen because of chance. In fact, under the null hypothesis
that the data only has a random relationship with my predictor the p-value
should have a uniform distribution across different data sets. So, if there is
no relationship you should expect to see all kinds of p-values and 80% is just
as plausible as 10%.

Therefore, I decided to check the distribution of p-values across data sets
(subjects). Specifically, I ran the cluster-level permutation analysis for all
subjects and stored the p-value of the maximum cluster in the data. If the
distribution over these p-values was uniform, there would be no relationship of
the support_correct predictor in my data. If the distribution was skewed towards
1, I would suspect an issue with the analysis and if the distribution was skewed
towards 0, I would interpret that as evidence that there is some evidence for
the relation between support_correct and MEG-signal in the data.

Results (p-values and identified clusters) are stored in

inf_results/clusres_5thdot_meg_201701171125.npz

The distribution of p-values is slightly skewed towards 0 with a median of 0.32.
The lowest p-values: 0.005, 0.008, 0.023, 0.024, 0.047 for subjects 34, 8, 7,
32, 18, respectively. So there is very weak, but existing evidence for a
relationship between support_correct and MEG-signal across subjects.

Technical issues
----------------
- Hame somehow forgot to put the raw fif-files for subject 15 into
  proni/BeeMEG/MEG/Raw/bm15a
- for subjects 14 and 30 there is behavioural data, but no MEG data, because
  Hame excluded their MEG data due to bad channels or eye movement artefacts