only check whether MEG signal supports correct or not
If the MEG signal is too complex and multi-faceted to directly show evidence for
a graded relationship to how much the 5th dot supports previous evidence, it
should be easier to find evidence for a simpler relationship. This relationship
could, e.g., simply be whether the 5th dot supports previous evidence or not,
instead of also looking for how much it supports it.

I could run this analysis very easily by just taking the sign of the
support_correct regressor. The results are stored in

clusres_5thdot_meg_201701171447.npz

As suspected, the p-value distribution became more skewed towards 0 indicating
more evidence for the relationship of support_correct to MEG-signal. The median
p-value was 0.18 and the 5 lowest p-values were

0.001, 0.001, 0.008, 0.012, 0.013 for subjects 33, 34, 17, 16, 18