clarifying use of 4-dot evidence, comparison to 'true' target
Main changes are in analyse_5thdot_behaviour.py where I now compare the analysis
with correct defined from the 4 previous dots vs. the traditional way of
defining it as the identity of the true underlying target that generated all
dots. To achieve this, trial_info returned by helpers.get_5th_dot_infos now
returns support_correct (based on true target) and support_correct_4th (based on
4-dot evidence).

Directly comparing the regression results with the two definitions of correct
shows that the main difference is an increase of p-values, when correct is
defined with the true targets, i.e., p-values tend to be between 5-10 orders of
magnitude higher, e.g., 10^-5 vs 10^-12, but across subjects the values are
strongly correlated (r=0.95). The same holds for the actual coefficients: They
tend to be smaller with correct defined by true targets, but the values are
highly correlated with those from defining correct using 4-dot evidence.

Overall this shows that the behavioural effect of the 5th dot shows much more
clearly in relation to evidence from previous dots than in relation to the true
targets.