dropping trials with RT < 550 ms
Based on the assumption that a trial with RT < 550 ms cannot contribute to any
effects of the 5th dot which was presnted from 500 to 600 ms I excluded these
trials from analysis.

There were not so many fast trials. Subject 37 had the most with 42 followed by
subject 24 with 29 fast trials. Most subjects had less than 10 fast trials.

I only checked the resulting p-values of the logistic regression and compared
these against those with all trials included. For 14 subjects p-values reduced,
when the fast trials were excluded. For 15 subjects p-values increased, but
across the 14 subjects the reductions in p-values were in sum larger
than the increases across the 15 subjects. This was because of 3 subjects which
had a particularly large drop in p-values compared with all other p-value
changes (reductions around or greater one log10-unit for subjects 10, 22, 24).

There was no clear relation between the number of fast trials that were excluded
for a subject and whether the p-value increased or decreased.

Excluding fast trials with RT < 600 ms lead to 17 subjects with increases in
p-values and 13 subjects with decreases. But the p-values for subjects 10, 22
and 24 decreased even further meaning that the sum of log10-p-value differences
was still larger for the subjects with p-value reductions.

Therefore, subjects 10, 22 and 24 appear to specifically benefit from exclusion
of fast trials while for most other subjects the reduction in number of trials
is not offset by larger power of support_correct to predict choices. Perhaps
there is still a beneficial effect in RTs for these subjects, though.

Note also that subjects 10, 22 and 24 were among the subjects with 'many' fast
trials, but there were other subjects with equally many fast trials (e.g., 37,
21, 32, 33, 11) which did not show these clear reductions in p-values.