computed dot-level posterior model measures for each subject
Computations take a while (30-60 min?) as they are not using the numba-code and
results still need to be checked. I use pandas' hdf-interface to store results.

I noticed that surprise and loglik_left are strongly correlated - need to look
into that.

Should I also include log-posterior odds as a measure? It shouldn't be very
informative additionally as the log-posterior already tells you which option is
winning and by how much. Anyway, you can compute log-posterior odds from the
log-posterior.