only consider values before a decision was made
The dot-level posterior measures were averaged over all simulated trials, but
should only be averaged over those trials in which the decision has not been
made yet at the specified moment in time. Actually there should only be a
difference for measures which are accumulated over time (log posterior values).
For the other measures, which only represent information from single dots,
removing trials in which a decision has already been made will only increase the
variance of the measure at the given time point. I should consider not removing
trials for these measures then.

The effect I expect is that the log-posterior values will asymptote to the bound
values with increasing time in a trial instead of steadily rising towards the
theoretical maximum of log(1)=0.