excluding averages with too few samples + 1st insights into measure correlations
To prevent extreme variation in measure averages resulting from averaging over
only very few samples I included new parameter 'Smin' which is the minimum
number of samples that need to be available for averaging, otherwise nan is
returned for that dot (time point) and trial.

I decided to keep the condition of averaging only over samples for which the
decision has not been reached yet also for the history-independent measures
(log-likelihoods, momentary evidence), because then the nan-ed times can at
least still indicate in this measure that it's very unlikely that this time
point will be reached in the corresponding trial.

I now also include logpost_right and loglik_right as these measures are not
necessarily simple inverses of the corresponding left-measures. The
log-posterior is much flatter for the winning model, than the loosing model (cf.
Bitzer2014) and the log-likelihoods should have slightly different behaviour
between the targets, as discussed below.

First insights into posterior measure correlations
--------------------------------------------------
Based on new measure values computed for all subjects.

Log-likelihoods
...............
The log-likelihoods for left and right targets are very strongly correlated
(~0.8). Reasons:
- unlikely dots (those with very large or very small x-values) tend to be
  unlikely with respect to both targets
- even when only looking at dots between the two targets where the
  log-likelihoods should differ most, the log-likelihoods are strongly
  correlated because the y-values of the dots contribute exactly the same amount
  of likelihood to the potentially different x-value-based likelihood, hence:
  the supposedly irrelevant y-values of the dots mask log-likelihood differences
  in the x-values

Surprise
........
Surprise is heavily anti-correlated with both log-likelihoods (<-0.9). This
correlation is even stronger than those between the log-likelihoods. The reason
is that surprise is something like a mixture of the two log-likehoods. The
mixture weights are the current posterior model probabilities. Because of that I
originally thought that surprise will differ from log-likelihoods, because it
has this history-dependent component, but this is completely overwhelmed by the
contribution from the log-likelihoods. Maybe this would not be the case, if the
log-likelihoods were not so strongly correlated themselves, but only as long as
the posterior model probabilities meander around 0.5, because as a model becomes
very likely surprise will only reflect the log-likelihood of that model.

Other than that surprise has the lowest correlations with all remaining measures
(<0.05).

Momentary evidence
..................
is moderately to strongly (~0.3 overall, but e.g., ~0.45 for subject 2)
correlated with the log-posterior values, indicating that individual dots may
strongly influence the accumulated evidence. Momentary evidence is also
moderately correlated with the log-likelihoods themselves (~0.3).

Log-posterior
.............
for the 'left' model is most strongly (anti-)correlated with its counterpart for
the right model (~-0.45), as expected. The relation is obviously very strong,
but nonlinear which is why the (linear) correlation is relatively low. This
correlation is mostly driven by values for which one log-posterior is close to 0
and the other is very negative.

There is a functional relationship between these two measures:
    logpost_2 = log(1 - exp(logpost_1))
When scatter plotted, all points lie below the corresponding line, but not
necessarily on the line. This is probably an effect of averaging the
log-posterior values independently from each other. The averages will be
strongly influenced by large negative values in the tails of the distribution
which pulls the averaged log-posterior values towards the lower left corner in
the scatter plot.

Is this additional randomness a bug or a feature? Do the scattered points
represent a genuine property of the estimated distribution of log-posterior
values across repetitions of the same trial? Or is this an unwanted artefact of
averaging the log-posterior for left and right independently?

Apart from the above reported correlation with momentary evidence, the
log-posterior values are only mildly correlated with the other measures.

Recommendation
..............
Based on my observations I should probably only include

- logpost_left
- m_evidence_left
- surprise

in subsequent regression analyses, where surprise should be seen as a
representation of the momentary log-likelihoods.