Extended cluster analysis with GLM
Having identified suitable regressors I have now coded a jitted function which
returns t-values as my earlier linregress_t, but this time for a full general
linear model. My changes in analyse_Hames_epochs.py demonstrate how this can be
used in cluster-based inference. The idea is that you select one regressor for
which you want to run the cluster-based inference. Two first tests with
regressors 'surprise' and 'logpost_left' for subject 3 gave no significant
clusters.

Note that my normalisations of the regressors on lines 173-175 do not affect the
estimated t-values for these regressors, but only the t-value of the intercept.