infer hierarchical GLM with ADVI across 5th-dot times and magnetometers
I have defined a hierarchical GLM across subjects using pyMC3. Specifically, I
allow different GLM parameters for each subject assuming that their values are
sampled from a Gaussian (random effects analysis). Of interest are then the
posterior mean and standard deviation of this Gaussian. Note that I fixed the
within-subject standard deviation to a common value across subjects.

I originally considered SPM's "summary statistic" approach which just means that
you find maximum likelihood estimates of GLM parameters per subject and then do
a t-test on the found parameter values. This approach is only valid under the
assumption that each subject has the same variance of the data which I find
rather doubtful. However, the approach is supposed to be quite robust so I might
cross-check against it later.

Summary measure
---------------
To get an overview of the effect I defined a statistic that is inspired by the
frequentist p-value: *the posterior p-value* which is the posterior probability
that the considered parameter takes on values with the opposite sign of the mean
value. In other words, if the posterior is calibrated, this is the probability
of erroneously thinking that the true parameter has the sign of the mean.
Another interpretation is that it is the probability of erroneously thinking
that the true parameter value is larger (or smaller for negative mean) than 0.

In my rush of getting results I was stupid enough to not save the posterior
distributions on which the posterior p-values were based.

Data
----
I let MNE export the magnetometer data to a DataFrame. MNE then automatically
multiplies the data with 1e15 which means that the channels get a standard
deviation of 100-250 units, a maximum absolute value of 1000-3500 units and a
mean around 0.

Priors
------
I pretty much stuck to the priors that were selected in the pyMC3 examples.
Each parameter mean had the same prior distribution (Gaussian with mean 0 and
std 100) which I assumed to be wide enough to cover all relevant parameter
values. Similarly, I assumed the same prior for the between-subject standard
deviations (Half-Cauchy with scale=5) and used the same prior for the
within-subject standard deviation.

Given that the different regressors have different scales (dot_y: 100,
surprise: 10, logpost_left: 5, entropy: 1, trial_time: 300, intercept: 1), I
wonder whether the chosen priors benefit the large-scaled regressors, because
they need smaller parameter values to reach similar effects which is easier to
reach with a restrictive prior penalising large values. Given that surprise and
logpost_left have similar scales, but very different 'posterior p-values' I
assume that the effect of the prior cannot be that large, but I should probably
simply normalise all data and regressors across subjects to exclude this
possibility altogether.

Results
-------
Topomaps of the posterior p-values for the 2nd-level mean parameters can be
found in the committed notebook. I plot the negative log10 transformed p-values,
i.e., the count of zeros after the period in the p-value. The strongest effect
is for surprise which exceeds values of 5 already at 400ms (5th dot onset), but
shows the strongest peaks with values around 30 and 20 at 550ms and 650ms. In
comparison, logpost_left has only a weak, but distinctive effect at 550-600ms
(values around 5). dot_y has effects with values around 10 at 600 and 700ms,
trial_time has a sustained left-lateralized effect with values in 5-10 from 500
to 700ms, entropy has its biggest effect (values >10) at 450ms and intercept has
no across-subject consistent effect (values < 2.1).

Practical issues
----------------
1. theano error when inverting diagonal matrix:
   I originally had a multivariate normal distribution with diagonal covariance
   matrix for the 1st-level parameters, but I couldn't run this model, because
   theano complained that the covariance matrix was positive definite and could
   not be inverted. I tried to resolve this by setting the testvalue of the
   between-subject standard deviation to non-zero which should have populated
   the diagonal of the covariance matrix properly, but the error remained. So I
   simply changed multivariate to a univariate normal which implements the same
   model. By the way, where are the posterior parameter correlations?
2. problems with file access:
   Because I originally saved only after the 102 channels for one time point
   were completed, the first run broke due to non-reachable proni
   (sftp-connection was gone). So I switched to saving after each channel, but
   at some point during the weekend the connection to proni was anyway broken. I
   expected this and put a try-statement around the call to save. Perhaps I
   should consider saving to local disk first.
3. memory fill:
   When I came back this morning the ipython console running my inference used
   15GB of RAM. I couldn't figure out why. Maybe this was due to some weird
   interaction with the non-reachable proni.